some minor bugs fixed

+++ MendezTransform/DotsImage.m
@@ -28,13 +28,13 @@ typedef struct dot_s {
     dots = _dots;
     dot_t   *dotset = (dot_t *)calloc(3 * self.dots, sizeof(dot_t));
     for (int i = 0; i < 3 * self.dots; i++) {
-        float   phi = 2 * M_PI * random() / (float)RAND_MAX;
-        float   theta = acos(1 - 2 * random() / (float)RAND_MAX);
+        float   phi = 2 * M_PI * (random() / (float)RAND_MAX);
+        float   theta = acos(1 - 2 * (random() / (float)RAND_MAX));
         float s = sin(theta);
         dotset[i].z = cos(theta);
         dotset[i].x = s * cos(phi);
         dotset[i].y = s * sin(phi);
-        dotset[i].r = 0.02 * (1 + 3 * random() / (float)RAND_MAX);
+        dotset[i].r = 0.02 * (1 + 3 * (random() / (float)RAND_MAX));
     }
     
     // create image
@@ -56,7 +56,7 @@ typedef struct dot_s {
             float   z = cosf(theta);
             unsigned char c[3] = { 0, 0, 0 };
             for (int i = 0; i < 3 * self.dots; i++) {
-                float a = acos(dotset[i].x * x + dotset[i].y * y + dotset[i].z * z);
+                float a = acosf(dotset[i].x * x + dotset[i].y * y + dotset[i].z * z);
                 if (a < dotset[i].r) {
                     int f = i % 6 + 1;
                     if (0x1 & f) { c[0] = 255; }
